✅ Create WebMVC

+++ src/Services/Aircraft/Aircrafts.API/Services/AircraftService.cs
@@ -1,6 +1,8 @@
 ﻿using Aircrafts.API.Repository;
 using AndreAirLines.Domain.Entities;
+using AndreAirLines.Domain.Entities.Enums;
 using AndreAirLines.Domain.Notifications;
+using AndreAirLines.Domain.Services;
 using AndreAirLines.Domain.Services.Base;
 using AndreAirLines.Domain.Validations;
 using System.Collections.Generic;
@@ -10,6 +12,7 @@ namespace Aircrafts.API.Services
 {
     public class AircraftService : BaseService
     {
+        private readonly GatewayService _gatewayService;
         private readonly IAircraftRepository _aircraftRepository;
 
         public AircraftService(IAircraftRepository aircraftRepository, INotifier notifier) : base(notifier)
@@ -27,18 +30,53 @@ namespace Aircrafts.API.Services
         {
             if (!ExecuteValidation(new AircraftValidation(), aircraft)) return aircraft;
 
+            var user = new User { LoginUser = aircraft.LoginUser };
+            await _gatewayService.PostLogAsync(user, null, aircraft, Operation.Create);
+
             return await _aircraftRepository.AddAsync(aircraft);
         }
 
         public async Task<Aircraft> UpdateAsync(Aircraft aircraft)
         {
+            var aircraftBefore = await _aircraftRepository.FindAsync(c => c.Id == aircraft.Id);
+
+
+            if (aircraftBefore == null)
+            {
+                Notification("Not found");
+                return aircraft;
+            }
+
+            var user = new User { LoginUser = aircraft.LoginUser };
+            await _gatewayService.PostLogAsync(user, aircraftBefore, aircraft, Operation.Update);
+
             return await _aircraftRepository.UpdateAsync(aircraft);
         }
 
-        public async Task RemoveAsync(Aircraft aircraftIn) =>
-            await _aircraftRepository.RemoveAsync(aircraftIn);
+        public async Task RemoveAsync(Aircraft aircraftIn)
+        {
+            var user = new User { LoginUser = aircraftIn.LoginUser };
+            await _gatewayService.PostLogAsync(user, aircraftIn, null, Operation.Delete);
 
-        public async Task RemoveAsync(string id) =>
+            await _aircraftRepository.RemoveAsync(aircraftIn);
+        }
+
+        public async Task<bool> RemoveAsync(string id)
+        {
+            var aircraft = await _aircraftRepository.FindAsync(c => c.Id == id);
+
+            if (aircraft == null)
+            {
+                Notification("Not found");
+                return false;
+            }
+
+            var user = new User { LoginUser = aircraft.LoginUser };
+            await _gatewayService.PostLogAsync(user, aircraft, null, Operation.Delete);
+
             await _aircraftRepository.RemoveAsync(id);
+
+            return true;
+        }
     }
 }
