+++ src/Services/Passenger/Passengers.API/Services/PassengerService.cs
@@ -1,4 +1,5 @@
 ﻿using AndreAirLines.Domain.Entities;
+using AndreAirLines.Domain.Entities.Enums;
 using AndreAirLines.Domain.Notifications;
 using AndreAirLines.Domain.Services;
 using AndreAirLines.Domain.Services.Base;
@@ -11,11 +12,13 @@ namespace Passengers.API.Services
 {
     public class PassengerService : BaseService
     {
+        private readonly GatewayService _gatewayService;
         private readonly IPassengerRepository _passengerRepository;
         private readonly ViaCepService _viaCepService;
 
-        public PassengerService(IPassengerRepository passengerRepository, ViaCepService viaCepService, INotifier notifier) : base(notifier)
+        public PassengerService(GatewayService gatewayService, IPassengerRepository passengerRepository, ViaCepService viaCepService, INotifier notifier) : base(notifier)
         {
+            _gatewayService = gatewayService;
             _passengerRepository = passengerRepository;
             _viaCepService = viaCepService;
         }
@@ -40,18 +43,54 @@ namespace Passengers.API.Services
 
             if (!ExecuteValidation(new PassengerValidation(), passenger)) return passenger;
 
+            var user = new User { LoginUser = passenger.LoginUser };
+            await _gatewayService.PostLogAsync(user, null, passenger, Operation.Create);
+
             return await _passengerRepository.AddAsync(passenger);
         }
 
         public async Task<Passenger> UpdateAsync(Passenger passenger)
         {
+
+            var passengerBefore = await _passengerRepository.FindAsync(c => c.Id == passenger.Id);
+
+
+            if (passengerBefore == null)
+            {
+                Notification("Not found");
+                return passenger;
+            }
+
+            var user = new User { LoginUser = passenger.LoginUser };
+            await _gatewayService.PostLogAsync(user, passengerBefore, passenger, Operation.Update);
+
             return await _passengerRepository.UpdateAsync(passenger);
         }
 
-        public async Task RemoveAsync(Passenger passengerIn) =>
-            await _passengerRepository.RemoveAsync(passengerIn);
+        public async Task RemoveAsync(Passenger passengerIn)
+        {
+            var user = new User { LoginUser = passengerIn.LoginUser };
+            await _gatewayService.PostLogAsync(user, passengerIn, null, Operation.Delete);
 
-        public async Task RemoveAsync(string id) =>
+            await _passengerRepository.RemoveAsync(passengerIn);
+        }
+
+        public async Task<bool> RemoveAsync(string id)
+        {
+            var passenger = await _passengerRepository.FindAsync(c => c.Id == id);
+
+            if (passenger == null)
+            {
+                Notification("Not found");
+                return false;
+            }
+
+            var user = new User { LoginUser = passenger.LoginUser };
+            await _gatewayService.PostLogAsync(user, passenger, null, Operation.Delete);
+
             await _passengerRepository.RemoveAsync(id);
+
+            return true;
+        }
     }
 }
